chore(otlp): rename module `expo_histogram` -> `histogram`
