+++ modules/observability-otlp/src/pipeline/metrics/histogram.rs
